pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "cell-operating-dataset"
version = "0.1.0"
description = "First production-oriented implementation of COD, the Cell Operating Dataset."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "pandas>=2.2",
  "pydantic>=2.7",
  "PyYAML>=6.0",
  "pyarrow>=15.0",
  "streamlit>=1.44",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
pythonpath = [
  "src",
]
testpaths = [
  "tests",
]
